Updated androidApp according to latest changes
