+++ 11-serialization/projects/starter/androidApp/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.raywenderlich.learn.android">
+    package="com.raywenderlich.learn">
 
     <uses-permission android:name="android.permission.INTERNET" />
