--- 11-serialization/projects/starter/androidApp/src/main/java/com/raywenderlich/learn/android/ui/utils/SingleLiveEvent.kt
@@ -1,91 +1,0 @@
-/*
- * Copyright (c) 2022 Razeware LLC
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * Notwithstanding the foregoing, you may not use, copy, modify, merge, publish,
- * distribute, sublicense, create a derivative work, and/or sell copies of the
- * Software in any work that is designed, intended, or marketed for pedagogical or
- * instructional purposes related to programming, coding, application development,
- * or information technology.  Permission for such use, copying, modification,
- * merger, publication, distribution, sublicensing, creation of derivative works,
- * or sale is expressly withheld.
- *
- * This project and source code may use libraries or frameworks that are
- * released under various Open-Source licenses. Use of those libraries and
- * frameworks are governed by their own individual licenses.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package com.raywenderlich.learn.android.ui.utils
-
-import androidx.annotation.MainThread
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.Observer
-import java.util.concurrent.atomic.AtomicBoolean
-import com.raywenderlich.learn.platform.Logger
-
-/**
- * A lifecycle-aware observable that sends only new updates after subscription, used for events like
- * navigation and Snackbar messages.
- *
- *
- * This avoids a common problem with events: on configuration change (like rotation) an update
- * can be emitted if the observer is active. This LiveData only calls the observable if there's an
- * explicit call to setValue() or call().
- *
- *
- * Note that only one observer is going to be notified of changes.
- */
-class SingleLiveEvent<T> : MutableLiveData<T>() {
-
-  private val pending = AtomicBoolean(false)
-
-  @MainThread
-  override fun observe(owner: LifecycleOwner, observer: Observer<in T>) {
-    if (hasActiveObservers()) {
-      Logger.w(TAG, "Multiple observers registered but only one will be notified of changes.")
-    }
-    // Observe the internal MutableLiveData
-    super.observe(owner, { t ->
-      if (pending.compareAndSet(true, false)) {
-        observer.onChanged(t)
-      }
-    })
-  }
-
-  @MainThread
-  override fun setValue(t: T?) {
-    pending.set(true)
-    super.setValue(t)
-  }
-
-  /**
-   * Used for cases where T is Void, to make calls cleaner.
-   */
-  @MainThread
-  fun call() {
-    value = null
-  }
-
-  companion object {
-
-    private const val TAG = "SingleLiveEvent"
-  }
-}
