+++ 11-serialization/projects/starter/androidApp/src/main/java/com/raywenderlich/learn/ui/main/MainScreen.kt
@@ -1,0 +1,130 @@
+/*
+ * Copyright (c) 2022 Razeware LLC
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * Notwithstanding the foregoing, you may not use, copy, modify, merge, publish,
+ * distribute, sublicense, create a derivative work, and/or sell copies of the
+ * Software in any work that is designed, intended, or marketed for pedagogical or
+ * instructional purposes related to programming, coding, application development,
+ * or information technology.  Permission for such use, copying, modification,
+ * merger, publication, distribution, sublicensing, creation of derivative works,
+ * or sale is expressly withheld.
+ *
+ * This project and source code may use libraries or frameworks that are
+ * released under various Open-Source licenses. Use of those libraries and
+ * frameworks are governed by their own individual licenses.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package com.raywenderlich.learn.ui.main
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.BottomSheetScaffold
+import androidx.compose.material.BottomSheetValue
+import androidx.compose.material.ExperimentalMaterialApi
+import androidx.compose.material.Scaffold
+import androidx.compose.material.rememberBottomSheetScaffoldState
+import androidx.compose.material.rememberBottomSheetState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.State
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.snapshots.SnapshotStateMap
+import androidx.compose.ui.unit.dp
+import androidx.navigation.compose.rememberNavController
+import com.raywenderlich.learn.ui.home.HomeSheetContent
+import com.raywenderlich.learn.ui.theme.colorContentSecondary
+import com.raywenderlich.learn.data.model.PLATFORM
+import com.raywenderlich.learn.data.model.RWEntry
+
+private lateinit var selected: MutableState<RWEntry>
+
+@OptIn(ExperimentalMaterialApi::class)
+@Composable
+fun MainScreen(
+  feeds: SnapshotStateMap<PLATFORM, List<RWEntry>>,
+  bookmarks: State<List<RWEntry>?>,
+  onUpdateBookmark: (RWEntry) -> Unit,
+  onShareAsLink: (RWEntry) -> Unit,
+  onOpenEntry: (String) -> Unit
+) {
+
+  val bottomNavigationItems = listOf(
+    BottomNavigationScreens.Home,
+    BottomNavigationScreens.Bookmark,
+    BottomNavigationScreens.Latest,
+    BottomNavigationScreens.Search
+  )
+
+  val navController = rememberNavController()
+  navController.enableOnBackPressed(false)
+
+  val coroutineScope = rememberCoroutineScope()
+  val bottomSheetScaffoldState = rememberBottomSheetScaffoldState(
+    bottomSheetState = rememberBottomSheetState(
+      initialValue = BottomSheetValue.Collapsed
+    )
+  )
+
+  selected = remember { mutableStateOf(RWEntry()) }
+
+  BottomSheetScaffold(
+    sheetContent = {
+      HomeSheetContent(
+        coroutineScope = coroutineScope,
+        bottomSheetScaffoldState = bottomSheetScaffoldState,
+        item = selected,
+        onUpdateBookmark = onUpdateBookmark,
+        onShareAsLink = onShareAsLink
+      )
+    },
+    sheetShape = RoundedCornerShape(
+      topStart = 16.dp,
+      topEnd = 16.dp
+    ),
+    sheetBackgroundColor = colorContentSecondary,
+    scaffoldState = bottomSheetScaffoldState, sheetPeekHeight = 0.dp
+  ) {
+
+    Scaffold(
+      topBar = {
+        MainTopAppBar()
+      },
+      bottomBar = {
+        MainBottomBar(
+          navController = navController,
+          items = bottomNavigationItems
+        )
+      },
+      content = {
+        MainContent(
+          navController = navController,
+          coroutineScope = coroutineScope,
+          bottomSheetScaffoldState = bottomSheetScaffoldState,
+          selected = selected,
+          feeds = feeds,
+          bookmarks = bookmarks,
+          onOpenEntry = onOpenEntry
+        )
+      }
+    )
+  }
+}
